requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-bootstrap4==22.3
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.0.0
Markdown==3.4.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
Pillow==9.4.0
protobuf==3.20.3
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
zipp==3.15.0
